Test property updates for Traits
These tests could've been done by modifying the `provideProperties()`
method and altering the templates based on the test it's providing for
which happens to be the first and only parameter* to providers methods

Such a "feature" would also require changing NOWDOC to HEREDOC though,
which in turn would mean escaping all the `$`s and making it unreadable.

* The parameter is entirely undocumented - I discovered it by calling
`func_get_args()` in the provider. It probably can't be relied on.