Add abstract methods for members & declarations
Rather than having an ugly switch, this lets the concrete classes define
the members and memberDeclarations themselves. Almost good to go now.

Would probably be better with an `openBrace` or `startPoint` method
returning `$node->classMembers->openBrace`, instead of the generic
`members` but It's only used once and I couldn't decide what to call it.