Always add property after traits & constants
I add this rule because when you create a new class and define a trait
before creating a constructor (for example).
When you create the constructor and you complete the attribute from it
(transformation action) then the properties were added before the trait
which was always a pain to correct manually.