Move the new line before property with phpdoc in the template
Before we made a decision that if a property has a type it will have a
phpdoc block because it was the only way to declare the type.

I don't like this because this logic depends on the templates and don't
have its place here anymore.
So I removed it and added the blank line directly in the template if
there is a phpdoc block.

Plus it fix the issue when inserting multiple properties for the first
time (no property in the class yet), then even if the properties have a
type there were no blank line between the properties.

Note: There is two cases where the blank line is a problem, when:
- the property is inserted after the class's open brace
- the is inserted after something else than the class's open brace or
another property