Removed Style fixer related classes
The time invested and still required in getting this system working, and
the complexity of ensuring that only generated code forces this to be
closed is too much to maintain.

Suggest that some of this can be retrieved in the future (e.g.
NodeQuery etc) to help in improving the existing updater approach.